Create file-selection handlers
