+++ index.html
@@ -57,5 +57,5 @@
 <script src="node_modules/quill/dist/quill.js"></script>
 <script src="editor.js"></script>
 <script>
-  require('./transcriptase')
+  require('./renderer')
 </script>
